made oracle run as ROSTest, added more launch scripts
