+++ src/wolf_control/scripts/controller.py
@@ -3,7 +3,8 @@ import rospy
 from geometry_msgs.msg import Vector3, Twist
 from std_msgs.msg import Float64
 
-class Controller:   
+
+class Controller:
     twist_state = Twist()
     yaw_control_out = 0.0
     depth_control_out = 0.0
@@ -25,7 +26,7 @@ class Controller:
     def __init__(self):
         rospy.init_node('controller', anonymous=False)
         rate = rospy.Rate(20)
-        
+
         rospy.Subscriber("wolf_imu_euler", Vector3, self.imu_callback)
         rospy.Subscriber("wolf_twist_setpoint", Twist, self.twist_callback)
         self.yaw_set_pub = rospy.Publisher("wolf_control/yaw_setpoint", Float64, queue_size=10)
@@ -41,6 +42,7 @@ class Controller:
             twist_out.angular.z = self.yaw_control_out
             self.twist_pub.publish(twist_out)
             rate.sleep()
+
 
 if __name__ == '__main__':
     controller = Controller()
